started redux
