+++ src/actions/todo.action.js
@@ -1,0 +1,3 @@
+function addTodo(value) {
+    return {type: "ADD_TODO", payload: value}
+}
